Queue now has a connect method. It is now enumerable and clearable. Color processor will try to get the color from the stick if no last colors are found.

+++ BlinkStickDotNet.Animations/Implementations/ChaseAnimation.cs
@@ -50,7 +50,7 @@ namespace BlinkStickDotNet.Animations.Implementations
                 offset = (offset + 1) % _colors.Length;
                 nr++;
 
-                Thread.Sleep(_duration / processor.NrOfLeds);
+                Thread.Sleep(_duration / (int)processor.NrOfLeds);
             }
         }
